chore: rename package and move details screen to v1 package
