+++ app/src/main/java/br/com/rstudio/countries/di/ApplicationModule.kt
@@ -27,15 +27,15 @@ import br.com.rstudio.countries.data.repository.QuizRepository
 import br.com.rstudio.countries.data.repository.QuizRepositoryImp
 import br.com.rstudio.countries.domain.GenerateQuizUseCase
 import br.com.rstudio.countries.domain.SaveQuizAnsweredUseCase
-import br.com.rstudio.countries.presentation.details.screen.DetailsContract
-import br.com.rstudio.countries.presentation.details.screen.DetailsPresenter
-import br.com.rstudio.countries.presentation.details.screen.DetailsTracker
 import br.com.rstudio.countries.presentation.homescreen.v1.ListContract
 import br.com.rstudio.countries.presentation.homescreen.v1.ListPresenter
 import br.com.rstudio.countries.presentation.homescreen.v1.ListTracker
 import br.com.rstudio.countries.presentation.homescreen.v2.HomeContract
 import br.com.rstudio.countries.presentation.homescreen.v2.HomePresenter
 import br.com.rstudio.countries.presentation.homescreen.v2.HomeTracker
+import br.com.rstudio.countries.presentation.overviewscreen.v1.DetailsContract
+import br.com.rstudio.countries.presentation.overviewscreen.v1.DetailsPresenter
+import br.com.rstudio.countries.presentation.overviewscreen.v1.DetailsTracker
 import br.com.rstudio.countries.presentation.quizscreen.QuizViewModel
 import com.bumptech.glide.Glide
 import com.bumptech.glide.load.engine.DiskCacheStrategy
