+++ app/src/main/java/br/com/rstudio/countries/presentation/homescreen/v2/HomeFragment.kt
@@ -16,7 +16,7 @@ import br.com.rstudio.countries.arch.extension.showError
 import br.com.rstudio.countries.arch.extension.showLoader
 import br.com.rstudio.countries.arch.model.ErrorModel
 import br.com.rstudio.countries.data.model.Country
-import br.com.rstudio.countries.presentation.details.screen.DetailsFragment
+import br.com.rstudio.countries.presentation.overviewscreen.v2.CountryOverviewFragment
 import br.com.rstudio.countries.presentation.homescreen.v1.model.ContinentVO
 import br.com.rstudio.countries.presentation.homescreen.v2.adapter.MainAdapter
 import org.koin.android.ext.android.inject
@@ -76,7 +76,7 @@ class HomeFragment : Fragment(R.layout.fragment_home), HomeContract.View {
   }
 
   override fun openDetailsScreen(country: Country, borders: List<Country>?) {
-    val fragment = DetailsFragment.newInstance(country, borders)
+    val fragment = CountryOverviewFragment.newInstance(country, borders)
     activity.replaceFragment(fragment)
   }
