+++ app/src/main/java/br/com/rstudio/countries/presentation/overviewscreen/v1/adapter/DetailsItemListViewHolder.kt
@@ -1,4 +1,4 @@
-package br.com.rstudio.countries.presentation.details.screen.adapter
+package br.com.rstudio.countries.presentation.overviewscreen.v1.adapter
 
 import android.view.View
 import android.view.ViewGroup
